Honestly this should've been broken up into smaller change sets, but fuck that apparently.
Hooray for more abstract layers of indirection.

Also the current design plan is to have generic PropertyModifier objects and potentially make it an interface to allow for more abstraction, then use the base Character object and generate a base CharacterState. Then the CharacterState object will have all of the PropertyModifiers applied to it with a priority system TBD.